2018-6-7 标定管理更加完善，同时修改所有页面按钮样式，批量查询功能优化，num从1开始，操作说明优化

+++ manageSystem/DAL/DBUtil.cs
@@ -25,7 +25,7 @@ namespace DAL
         private void CreateToolsInfoDb()
         {
             string[] toolsInfoFeildName = new string[] { "SerialNum", "Model", "Category", "Name", "TorqueMin", "TorqueMax", "Accuracy", "Section", "Workstation", "DemarcateCycle", "Status", "QualityAssureDate", "MaintainContractStyle", "MaintainContractDate", "RepairTimes", "Remark" };
-            string[] toolsInfoFeildType = new string[] { "VARCHAR(255) PRIMARY KEY", "VARCHAR(255)", "VARCHAR(255)", "VARCHAR(255)", "Integer", "Integer", "Integer", "VARCHAR(255)", "VARCHAR(255)", "Integer", "VARCHAR(255)", "VARCHAR(255)", "VARCHAR(255)" , "VARCHAR(255)", "Integer", "VARCHAR(255)" };
+            string[] toolsInfoFeildType = new string[] { "VARCHAR(255) PRIMARY KEY", "VARCHAR(255)", "VARCHAR(255)", "VARCHAR(255)", "DOUBLE", "DOUBLE", "DOUBLE", "VARCHAR(255)", "VARCHAR(255)", "Integer", "VARCHAR(255)", "VARCHAR(255)", "VARCHAR(255)" , "VARCHAR(255)", "Integer", "VARCHAR(255)" };
             SQLHelper.CreateTable("ToolsInfo", toolsInfoFeildName, toolsInfoFeildType);
         }
